# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DOCUMENT RWANDA",
        "version": "1.0"
    },
    "paths": {
...
        "/api/nationality/list/{lg}": {
            "get": {
                "tags": [
                    "Lists"
                ],
                "description": "Liste de Nationalités.",
                "operationId": "nationality",
                "responses": {
                    "200": {
                        "description": "Liste de Nationalités."
                    },
                    "400": {
                        "description": "Bad Request."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                }
            }
        },
        "/api/country/list/{lg}": {
            "get": {
                "tags": [
                    "Lists"
                ],
                "description": "Liste de Pays.",
                "operationId": "country",
                "responses": {
                    "200": {
                        "description": "Liste de Pays."
                    },
                    "400": {
                        "description": "Bad Request."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                }
            }
        },
        "/api/provinces/list/{lg}": {
            "get": {
                "tags": [
                    "Lists"
                ],
                "description": "Liste de provinces.",
                "operationId": "province",
                "responses": {
                    "200": {
                        "description": "Liste de provinces."
                    },
                    "400": {
                        "description": "Bad Request."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                }
            }
        },
...
        "/api/districts/list/{lg}/{province_id}": {
            "get": {
                "tags": [
                    "Lists"
                ],
                "description": "Liste de districts.",
                "operationId": "district",
                "responses": {
                    "200": {
                        "description": "Liste de districts."
                    },
                    "400": {
                        "description": "Bad Request."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                }
            }
        },
...
        "/api/spouses": {
            "get": {
                "tags": [
                    "Spouses"
                ],
                "description": "Liste des Conjoints",
                "operationId": "listSpouse",
                "responses": {
                    "200": {
                        "description": "Liste des Conjoints."
                    },
                    "400": {
                        "description": "Bad Request."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/spouses/{uid}": {
            "get": {
                "tags": [
                    "Spouses"
                ],
                "description": "Détail du conjoint",
                "operationId": "showSpouse",
                "responses": {
                    "200": {
                        "description": "Détail du conjoint."
                    },
                    "400": {
                        "description": "Bad Request."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Spouses"
                ],
                "description": "Modification d'un conjoint",
                "operationId": "editSpouse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lastname",
                                    "firstname",
                                    "gender",
                                    "number",
                                    "email",
                                    "birthday",
                                    "birthplace",
                                    "profession",
                                    "village",
                                    "street_number",
                                    "house_number",
                                    "family_number",
                                    "fullname_person",
                                    "number_person",
                                    "residence_person",
                                    "cellule_id",
                                    "maritalstatus_id",
                                    "town_id",
                                    "fullname_father",
                                    "fullname_mother",
                                    "status",
                                    "profile_id"
                                ],
                                "properties": {
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "birthplace": {
                                        "type": "string"
                                    },
                                    "profession": {
                                        "type": "string"
                                    },
                                    "village": {
                                        "type": "string"
                                    },
                                    "street_number": {
                                        "type": "string"
                                    },
                                    "house_number": {
                                        "type": "string"
                                    },
                                    "family_number": {
                                        "type": "integer"
                                    },
                                    "register_number": {
                                        "type": "integer"
                                    },
                                    "fullname_person": {
                                        "type": "string"
                                    },
                                    "number_person": {
                                        "type": "string"
                                    },
                                    "residence_person": {
                                        "type": "string"
                                    },
                                    "cellule_id": {
                                        "type": "integer"
                                    },
                                    "town_id": {
                                        "type": "integer"
                                    },
                                    "bp": {
                                        "type": "string"
                                    },
                                    "diplome": {
                                        "type": "string"
                                    },
                                    "distinction": {
                                        "type": "string"
                                    },
                                    "maritalstatus_id": {
                                        "type": "integer"
                                    },
                                    "nationality_id": {
                                        "type": "integer"
                                    },
                                    "fullname_father": {
                                        "type": "string"
                                    },
                                    "fullname_mother": {
                                        "type": "string"
                                    },
                                    "comment": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "integer"
                                    },
                                    "profile_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "conjoint modifié avec succès."
                    },
                    "400": {
                        "description": "Erreur de validation."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
}